Langchain/vector/requirements.txt
langchain-chroma
chromadb
langchain-google-genai
langchain
python-dotenv
google-genai
pydantic
langchain-core
langchain-huggingface
youtube_transcript_api
faiss-cpu